Move AdditionalRegistrationFields from to models.py
While running the tests Open edX creates database based on models
in models.py files of all of the apps. Putting this model in models.py
fixes the failing tests.
